Moved instructor files components into project admin folder.
