+++ src/components/project_admin/instructor_files/single_instructor_file.vue
@@ -73,12 +73,12 @@
 </template>
 
 <script lang="ts">
-  import APIErrors from '@/components/api_errors.vue';
-  import Modal from '@/components/modal.vue';
-  import ValidatedForm from '@/components/validated_form.vue';
-  import ValidatedInput from '@/components/validated_input.vue';
-  import { handle_api_errors_async } from '@/utils';
-  import { is_not_empty } from '@/validators';
+  import APIErrors from '../../api_errors.vue';
+  import Modal from '../../modal.vue';
+  import ValidatedForm from '../../validated_form.vue';
+  import ValidatedInput from '../../validated_input.vue';
+  import { handle_api_errors_async } from '../../../utils';
+  import { is_not_empty } from '../../../validators';
   import { InstructorFile } from 'ag-client-typescript';
   import * as FileSaver from 'file-saver';
